新增家庭户审核、股权变更审核、平台内通知
审核模块化，平台内通知

+++ src/main/java/zyxhj/jiti/domain/ORGLoginBo.java
@@ -36,7 +36,7 @@ public class ORGLoginBo {
 	public String shareCerImg;
 	public Boolean shareCerHolder;
 
-	public Integer shareAmount;
+	public Double shareAmount;
 	public Integer weight;
 
 	public JSONArray orgRoles;
